Remade order controller into service and controller And fixed small issues on client
