+++ ecommerce-api/ecommerce-api/src/controllers/orderController.ts
@@ -1,42 +1,19 @@
-import { Request, Response } from "express";
+import { json, Request, Response } from "express";
 import { db } from "../config/db";
 import { IOrderItem } from "../models/IOrderItem";
 import { IOrder } from "../models/IOrder";
 import { logError } from "../utilities/logger";
 import { ResultSetHeader } from "mysql2";
 
+import * as orderService from "../services/orderService";
+import { formatOrderDetails } from "../utilities/dataHandler";
+
 export const getOrders = async (_: any, res: Response) => {
   try {
-    const sql = `
-      SELECT 
-        orders.id AS id, 
-        customer_id,
-        total_price,
-        payment_status,
-        payment_id,
-        order_status,
-        orders.created_at,
-        customers.id AS customer_id, 
-        firstname AS customer_firstname, 
-        lastname AS customer_lastname, 
-        email AS customer_email, 
-        firstname AS customer_firstname, 
-        lastname AS customer_lastname,
-        email AS customer_email,
-        phone AS customer_phone,
-        street_address AS customer_street_address,
-        postal_code AS customer_postal_code,
-        city AS customer_city,
-        country AS customer_country,
-        customers.created_at AS customers_created_at,
-        orders.created_at AS created_at
-      
-      FROM orders
-      LEFT JOIN customers ON orders.customer_id = customers.id`;
-    const [rows] = await db.query<IOrder[]>(sql);
-    res.json(rows);
+    const orders = await orderService.getOrders();
+    res.status(200).json(orders);
   } catch (error) {
-    res.status(500).json({ error: logError(error) });
+    res.status(500).json({ message: logError(error) });
   }
 };
 
@@ -44,22 +21,10 @@ export const getOrderById = async (req: Request, res: Response) => {
   const id: string = req.params.id;
 
   try {
-    const sql = `
-      SELECT 
-        *, 
-        orders.created_at AS orders_created_at, 
-        customers.created_at AS customers_created_at 
-      FROM orders 
-      LEFT JOIN customers ON orders.customer_id = customers.id
-      LEFT JOIN order_items ON orders.id = order_items.order_id
-      WHERE orders.id = ?
-    `;
-    const [rows] = await db.query<IOrder[]>(sql, [id]);
-    // res.json(rows)
-    // return;
+    const order = await orderService.getOrderById(Number(id));
 
-    rows && rows.length > 0
-      ? res.json(formatOrderDetails(id, rows))
+    order
+      ? res.status(200).json(order)
       : res.status(404).json({ message: "Order not found" });
   } catch (error) {
     res.status(500).json({ error: logError(error) });
@@ -67,132 +32,37 @@ export const getOrderById = async (req: Request, res: Response) => {
 };
 
 export const getOrderByPaymentID = async (req: Request, res: Response) => {
-  const id: string = req.params.payment_id;
-
-  console.log(id);
+  const payment_id: string = req.params.payment_id;
 
   try {
-    const sql = `
-      SELECT 
-        *, 
-        orders.created_at AS orders_created_at, 
-        customers.created_at AS customers_created_at 
-      FROM orders 
-      LEFT JOIN customers ON orders.customer_id = customers.id
-      LEFT JOIN order_items ON orders.id = order_items.order_id
-      WHERE orders.payment_id = ?
-    `;
-    const [rows] = await db.query<IOrder[]>(sql, [id]);
-    // res.json(rows)
-    // return;
+    const order = await orderService.getOrderByPaymentId(payment_id);
 
-    rows && rows.length > 0
-      ? res.json(formatOrderDetails(id, rows))
+    order
+      ? res.status(200).json(order)
       : res.status(404).json({ message: "Order not found" });
   } catch (error) {
     res.status(500).json({ error: logError(error) });
   }
 };
 
-const formatOrderDetails = (orderId, rows) => ({
-  id: orderId,
-  customer_id: rows[0].customer_id,
-  total_price: rows[0].total_price,
-  payment_status: rows[0].payment_status,
-  payment_id: rows[0].payment_id,
-  order_status: rows[0].order_status,
-  created_at: rows[0].orders_created_at,
-  customer_firstname: rows[0].firstname,
-  customer_lastname: rows[0].lastname,
-  customer_email: rows[0].email,
-  customer_password: rows[0].password,
-  customer_phone: rows[0].phone,
-  customer_street_address: rows[0].street_address,
-  customer_postal_code: rows[0].postal_code,
-  customer_city: rows[0].city,
-  customer_country: rows[0].country,
-  order_items: rows[0].id
-    ? rows.map((item: IOrderItem) => ({
-        id: item.id,
-        product_id: item.product_id,
-        product_name: item.product_name,
-        quantity: item.quantity,
-        unit_price: item.unit_price,
-        image_url: item.image_url,
-      }))
-    : [],
-});
-
 export const createOrder = async (req: Request, res: Response) => {
   const { customer_id, payment_status, payment_id, order_status }: IOrder =
     req.body;
 
   try {
-    const sql = `
-      INSERT INTO orders (customer_id, total_price, payment_status, payment_id, order_status)
-      VALUES (?, ?, ?, ?, ?)
-    `;
-
-    const totalPrice = req.body.order_items.reduce(
-      (total, item) => total + item.quantity * item.unit_price,
-      0
-    );
-    const params = [
+    const createdOrderID = await orderService.createOrder(
       customer_id,
-      totalPrice,
       payment_status,
       payment_id,
       order_status,
-    ];
-    const [result] = await db.query<ResultSetHeader>(sql, params);
-    if (result.insertId) {
-      const order_id: number = result.insertId;
-      const orderItems = req.body.order_items;
-      for (const orderItem of orderItems) {
-        const data = { ...orderItem, order_id };
-        await createOrderItem(data);
-      }
-    }
+      req.body.order_items
+    );
 
     res
       .status(201)
-      .json({ message: "Order created", insertedID: result.insertId });
+      .json({ message: "Order created", insertedID: createdOrderID });
   } catch (error: unknown) {
     res.status(500).json({ error: logError(error) });
-  }
-};
-
-const createOrderItem = async (data: IOrderItem) => {
-  const {
-    order_id,
-    product_id,
-    product_name,
-    quantity,
-    unit_price,
-    image_url,
-  } = data;
-  try {
-    const sql = `
-      INSERT INTO order_items (
-        order_id, 
-        product_id, 
-        product_name, 
-        quantity, 
-        unit_price,
-        image_url 
-      ) VALUES (?, ?, ?, ?, ?, ?)
-    `;
-    const params = [
-      order_id,
-      product_id,
-      product_name,
-      quantity,
-      unit_price,
-      image_url,
-    ];
-    await db.query<ResultSetHeader>(sql, params);
-  } catch (error) {
-    throw new Error();
   }
 };
 
@@ -201,17 +71,16 @@ export const updateOrder = async (req: Request, res: Response) => {
   const { payment_status, payment_id, order_status }: IOrder = req.body;
 
   try {
-    const sql = `
-      UPDATE orders 
-      SET payment_status = ?, payment_id = ?,order_status = ?
-      WHERE id = ?
-    `;
-    const params = [payment_status, payment_id, order_status, id];
-    const [result] = await db.query<ResultSetHeader>(sql, params);
+    const updatedOrder = await orderService.updateOrder(
+      Number(id),
+      payment_status,
+      payment_id,
+      order_status
+    );
 
-    result.affectedRows === 0
-      ? res.status(404).json({ message: "Order not found" })
-      : res.json({ message: "Order updated" });
+    updatedOrder
+      ? res.status(200).json({ message: "Order updated" })
+      : res.status(404).json({ message: "Order not found" });
   } catch (error) {
     res.status(500).json({ error: logError(error) });
   }
@@ -221,12 +90,11 @@ export const deleteOrder = async (req: Request, res: Response) => {
   const id: string = req.params.id;
 
   try {
-    const sql = "DELETE FROM orders WHERE id = ?";
-    const [result] = await db.query<ResultSetHeader>(sql, [id]);
+    const deletedOrder = await orderService.deleteOrder(Number(id));
 
-    result.affectedRows === 0
-      ? res.status(404).json({ message: "Order not found" })
-      : res.json({ message: "Order deleted" });
+    deletedOrder
+      ? res.status(204)
+      : res.status(404).json({ message: "Ordeer not found" });
   } catch (error) {
     res.status(500).json({ error: logError(error) });
   }
